New home component to handle navigation reset
